// src/app/Rest/Conductor/model/taskDef.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TaskDef { 
    ownerApp?: string;
    createTime?: number;
    updateTime?: number;
    createdBy?: string;
    updatedBy?: string;
    name: string;
    description?: string;
    retryCount?: number;
    timeoutSeconds: number;
    inputKeys?: Array<string>;
    outputKeys?: Array<string>;
    timeoutPolicy?: TaskDef.TimeoutPolicyEnum;
    retryLogic?: TaskDef.RetryLogicEnum;
    retryDelaySeconds?: number;
    responseTimeoutSeconds?: number;
    concurrentExecLimit?: number;
    inputTemplate?: { [key: string]: any; };
    rateLimitPerFrequency?: number;
    rateLimitFrequencyInSeconds?: number;
    isolationGroupId?: string;
    executionNameSpace?: string;
    ownerEmail?: string;
    pollTimeoutSeconds?: number;
}
export namespace TaskDef {
    export type TimeoutPolicyEnum = 'RETRY' | 'TIME_OUT_WF' | 'ALERT_ONLY';
    export const TimeoutPolicyEnum = {
        RETRY: 'RETRY' as TimeoutPolicyEnum,
        TIMEOUTWF: 'TIME_OUT_WF' as TimeoutPolicyEnum,
        ALERTONLY: 'ALERT_ONLY' as TimeoutPolicyEnum
    };
    export type RetryLogicEnum = 'FIXED' | 'EXPONENTIAL_BACKOFF';
    export const RetryLogicEnum = {
        FIXED: 'FIXED' as RetryLogicEnum,
        EXPONENTIALBACKOFF: 'EXPONENTIAL_BACKOFF' as RetryLogicEnum
    };
}